Kill entire process group when killing process
Previously when we would send a Kill command to a subprocess, we would
not send that same command to processes started by the subprocess (the
grandchildren, if you will). This resulted in goreman hanging when you
send a ctrl+C to it.

Sending -1*pid in the syscall tells Unix to kill the entire process
group - the child and the grandchildren, not just the child.

Verified that this fixes the hang I observed when running
dev/launch.sh manually.

For more, see
https://medium.com/@felixge/killing-a-child-process-and-all-of-its-children-in-go-54079af94773.
I am not sure there is a Windows analog so for that we continue to use
cmd.Process.Kill().