Add flag to avoid setting ports
Some configurations can't handle dynamic port setting, rely on
hard coding ports, and/or don't need a port. In these instances the
use of "Starting X on port <Y>" is confusing when the port is not
actually being set for this value.

Add a new --set-ports flag which you can set to False to disable port
setting, and the subsequent logs.